@@ -32,19 +32,20 @@
     "index": "docker run -it --env-file=.env -e \"CONFIG=$(cat docsearch.json | jq -r tostring)\" algolia/docsearch-scraper"
   },
   "dependencies": {
-    "@docusaurus/core": "3.0.0-alpha.0",
-    "@docusaurus/plugin-ideal-image": "3.0.0-alpha.0",
-    "@docusaurus/plugin-pwa": "3.0.0-alpha.0",
-    "@docusaurus/preset-classic": "3.0.0-alpha.0",
-    "@docusaurus/theme-search-algolia": "3.0.0-alpha.0",
+    "@docusaurus/core": "3.0.0",
+    "@docusaurus/plugin-ideal-image": "3.0.0",
+    "@docusaurus/plugin-pwa": "3.0.0",
+    "@docusaurus/preset-classic": "3.0.0",
+    "@docusaurus/theme-search-algolia": "3.0.0",
     "@giscus/react": "^2.3.0",
     "@popperjs/core": "^2.11.8",
-    "dayjs": "^1.11.9",
+    "dayjs": "^1.11.10",
     "docusaurus-plugin-image-zoom": "^0.1.4",
     "docusaurus-plugin-sass": "^0.2.5",
-    "framer-motion": "^10.13.1",
+    "framer-motion": "^10.16.4",
     "loadsh": "^0.0.4",
     "ora": "^5.4.1",
+    "prism-react-renderer": "^2.1.0",
     "raw-loader": "^4.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -52,14 +53,15 @@
     "sass": "^1.64.1"
   },
   "devDependencies": {
-    "@docusaurus/eslint-plugin": "3.0.0-alpha.0",
-    "@docusaurus/module-type-aliases": "3.0.0-alpha.0",
+    "@docusaurus/eslint-plugin": "3.0.0",
+    "@docusaurus/module-type-aliases": "3.0.0",
+    "@docusaurus/tsconfig": "^3.0.0",
+    "@docusaurus/types": "3.0.0",
     "@iconify/react": "^4.1.1",
-    "@tsconfig/docusaurus": "^1.0.7",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
-    "eslint": "^8.45.0",
-    "eslint-config-prettier": "^8.8.0",
+    "eslint": "^8.53.0",
+    "eslint-config-prettier": "^9.0.09.0.0",
     "eslint-plugin-prettier": "^4.2.1",
     "prettier": "^2.8.8",
     "stylelint": "^14.16.1",
@@ -67,18 +69,9 @@
     "stylelint-config-prettier-scss": "^0.0.1",
     "stylelint-config-standard": "^26.0.0",
     "stylelint-config-standard-scss": "^5.0.0",
-    "typescript": "^4.9.5"
+    "typescript": "~5.2.2"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "engines": {
+    "node": ">=18.0"
   }
 }
